fsperf: Allow same block address in the read and write pending I/O queue
Now, fsperf stops to run if it found the same block address in pending
I/O queue. Because FS usually should not issue same block address for
read and write.

But, on tux3, for now, it issues same block address for read and write.
[Fix is planned, but it may not be done soon. Because, in the case of
tux3, it is just a performance issue, not correctness issue.]

So, this allows above situation (same block address for read and
write) by "--no-error" option.

To do it, this separates pending queue to read and write. Then, it
adds sanity check for same block address. If sanity check found same
address, it stop to run if "--no-error" option was not specified.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>